#15 REFACTOR renames, autoformating
