+++ src/scraper/builder.py
@@ -4,7 +4,7 @@ Created on 18.05.2018
 @author: Mirco
 '''
 from scraper.scraper_bloomberg import ScraperBloomberg
-from scraper.stocknames_importer_csv import ImporterCSV
+from scraper.importer_csv import ImporterCSV
 from scraper.manager import Manager
 from scraper.security_list import SecurityList
 from scraper.exporter_csv import Exporter_CSV
